LastBootupTime
This script will retrieve the last boot time for each server in your list and export the data to a CSV file while maintaining the computer name (CSName) and the last boot time (LastBootUpTime) in the output file. Make sure to replace the file paths with your actual paths if they differ.